Worker interface changed, fn as a parameter design deleted from worker
