--- packages/js/src/container/index.ts
@@ -1,2 +1,0 @@
-export * from '../account/account';
-export * from './runtime';
