+++ packages/js/src/worker.ts
@@ -1,0 +1,128 @@
+import {NEAR} from 'near-units';
+import {getNetworkFromEnv, urlConfigFromNetwork} from './utils';
+import {ClientConfig} from './types';
+import {AccountManager, NearAccount, NearAccountManager} from './account';
+import {Config} from './interfaces';
+import {JsonRpcProvider} from './jsonrpc';
+import {debug} from './internal-utils';
+import {SandboxServer} from './server/server';
+
+export abstract class Worker {
+  protected config: Config;
+
+  protected manager!: NearAccountManager;
+
+  constructor(config: Config) {
+    debug('Lifecycle.Worker.constructor', 'config:', config);
+    this.config = config;
+    this.manager = AccountManager.create(config);
+  }
+
+  static async init(config: Partial<Config> = {}): Promise<Worker> {
+    debug('Lifecycle.Worker.init()', 'config:', config);
+    switch (config.network ?? getNetworkFromEnv()) {
+      case 'testnet':
+        return TestnetWorker.init(config);
+      case 'sandbox':
+        return SandboxWorker.init(config);
+      default:
+        throw new Error(
+          `config.network = '${config.network}' invalid; ` // eslint-disable-line @typescript-eslint/restrict-template-expressions
+            + 'must be \'testnet\' or \'sandbox\' (the default). Soon \'mainnet\'',
+        );
+    }
+  }
+
+  get rootAccount(): NearAccount {
+    return this.manager.root;
+  }
+
+  abstract get provider(): JsonRpcProvider;
+
+  abstract tearDown(): Promise<void>;
+}
+
+export class TestnetWorker extends Worker {
+  static async init(config: Partial<Config>): Promise<TestnetWorker> {
+    debug('Lifecycle.TestnetWorker.create()', 'config:', config);
+    const fullConfig = {...this.defaultConfig, ...config};
+    const runtime = new TestnetWorker(fullConfig);
+    await runtime.manager.init();
+    return runtime;
+  }
+
+  get provider(): JsonRpcProvider {
+    return JsonRpcProvider.from(TestnetWorker.clientConfig);
+  }
+
+  async tearDown(): Promise<void> {
+    // We are not stoping any server here because we are using Testnet
+    return Promise.resolve();
+  }
+
+  private static get defaultConfig(): Config {
+    return {
+      homeDir: 'ignored',
+      port: 3030,
+      init: true,
+      rm: false,
+      refDir: null,
+      ...this.clientConfig,
+    };
+  }
+
+  private static get clientConfig(): ClientConfig {
+    return urlConfigFromNetwork('testnet');
+  }
+}
+
+export class SandboxWorker extends Worker {
+  private server!: SandboxServer;
+
+  static async init(config: Partial<Config>): Promise<SandboxWorker> {
+    debug('Lifecycle.SandboxWorker.create()', 'config:', config);
+    const defaultConfig = await this.defaultConfig();
+    const worker = new SandboxWorker({...defaultConfig, ...config});
+    worker.server = await SandboxServer.init(worker.config);
+    await worker.server.start();
+    return worker;
+  }
+
+  private static async defaultConfig(): Promise<Config> {
+    const port = await SandboxServer.nextPort();
+    return {
+      ...this.clientConfig,
+      homeDir: SandboxServer.randomHomeDir(),
+      port,
+      init: true,
+      rm: false,
+      refDir: null,
+      rpcAddr: `http://localhost:${port}`,
+    };
+  }
+
+  get provider(): JsonRpcProvider {
+    return JsonRpcProvider.from(this.rpcAddr);
+  }
+
+  async tearDown(): Promise<void> {
+    try {
+      await this.server.close();
+    } catch (error: unknown) {
+      debug('this.server.close() threw error.', JSON.stringify(error, null, 2));
+    }
+  }
+
+  private static get clientConfig(): ClientConfig {
+    return {
+      network: 'sandbox',
+      rootAccount: 'test.near',
+      rpcAddr: '', // Will be over written
+      initialBalance: NEAR.parse('100 N').toJSON(),
+    };
+  }
+
+  private get rpcAddr(): string {
+    return `http://localhost:${this.config.port}`;
+  }
+}
